--- lib/colors.dart
@@ -1,19 +1,0 @@
-import 'package:flutter/material.dart';
-
-/// COLORS FILE
-/// It contains app colors to used them all around the app
-const primaryColor = const Color(0xFF151515);
-const accentColor = const Color(0xFFFFBC02D);
-const backgroundColor = const Color(0xFF212121);
-const cardColor = const Color(0xFF2C2C2C);
-
-const primaryText = const Color(0xFFFFFFFF);
-const secondaryText = const Color(0xFFBCBCBC);
-const lateralText = const Color(0xFF828282);
-
-const acceptIcon = const Color(0xFF4CAF50);
-const denyIcon = const Color(0xFFF44336);
-const nullIcon = const Color(0xFF607D8B);
-
-const highlightColor = const Color(0xFF424242);
-const dividerColor = const Color(0xFF616161);
