--- lib/views/home_page.dart
@@ -1,82 +1,0 @@
-import 'package:flutter/material.dart';
-
-import '../url.dart';
-import 'about_page.dart';
-import 'launch_list.dart';
-import 'vehicle_list.dart';
-
-/// HOME PAGE CLASS
-/// Home page of the app.
-class HomePage extends StatefulWidget {
-  @override
-  State<StatefulWidget> createState() => _HomePageState();
-}
-
-class _HomePageState extends State<HomePage>
-    with SingleTickerProviderStateMixin {
-  TabController _tabController;
-  List<StatelessWidget> _homeLists = List(_tabs.length);
-
-  /// List of the TabBar's tabs
-  static const List<Tab> _tabs = const <Tab>[
-    const Tab(text: 'VEHICLES'),
-    const Tab(text: 'UPCOMING'),
-    const Tab(text: 'LATEST'),
-  ];
-
-  @override
-  void initState() {
-    super.initState();
-
-    // Tab controller init
-    _tabController = TabController(
-      length: _tabs.length,
-      vsync: this,
-      initialIndex: 1,
-    );
-
-    // List array init
-    _homeLists = [
-      VehicleList(),
-      LaunchList(Url.upcomingList),
-      LaunchList(Url.launchesList),
-    ];
-  }
-
-  @override
-  void dispose() {
-    _tabController.dispose();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text(
-          'SpaceX GO!',
-          style: TextStyle(fontWeight: FontWeight.bold),
-        ),
-        centerTitle: true,
-        actions: <Widget>[
-          IconButton(
-            icon: const Icon(Icons.info_outline),
-            onPressed: () => Navigator.push(
-                  context,
-                  MaterialPageRoute(builder: (context) => AboutPage()),
-                ),
-          ),
-        ],
-        bottom: TabBar(
-          labelStyle: TextStyle(
-            fontFamily: 'ProductSans',
-            fontWeight: FontWeight.bold,
-          ),
-          controller: _tabController,
-          tabs: _tabs,
-        ),
-      ),
-      body: TabBarView(controller: _tabController, children: _homeLists),
-    );
-  }
-}
